refactor: move board methods to separate service
